AI_Assistant/requirements.txt
numpy
pandas
ipykernel
sentence_transformers
faiss-cpu
twilio